% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_funcs.R
\name{acceptRate}
\alias{acceptRate}
\title{A acceptance rate function.}
\usage{
acceptRate(df, burn = 0)
}
\arguments{
\item{df}{A data frame of samples. Rows correspond with iterations, and columns with parameters.}

\item{burn}{Number of rows you want to discard as burn in. Defaults to 0.}
}
\description{
This function calculates the acceptance rate for samples stored in a data frame.
}
\examples{
fake_samples <- data.frame(rnorm(100))
acceptRate(fake_samples)
}
\keyword{accept}
\keyword{acceptance}
\keyword{rate}
