% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_funcs.R
\name{correlogram}
\alias{correlogram}
\title{A correlogram function.}
\usage{
correlogram(df, cols, burn = 0, ...)
}
\arguments{
\item{df}{A data frame of samples. Rows correspond with iterations, and columns with parameters.}

\item{cols}{column indices of the parameters you're interested in plotting.}

\item{burn}{Number of rows you want to discard as burn in. Defaults to 0.}

\item{...}{Extra arguments to be passed into stats::acf() (e.g. lag.max )}
}
\description{
This function plots overlapping correlograms for samples stored in a data frame.
}
\examples{
fake_samples <- data.frame(rnorm(100))
correlogram(fake_samples, 1, lag.max = 100)
}
\keyword{acf}
\keyword{autocorrelation}
\keyword{correlogram}
