% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_funcs.R
\name{elicitBeta}
\alias{elicitBeta}
\title{A function that helps you elicits beta prior parameters.}
\usage{
elicitBeta(ideal_mean, percent_ci, lower_ci, upper_ci, beta_grid_left,
  beta_grid_right, beta_increment, plot = T)
}
\arguments{
\item{ideal_mean}{desired mean for your random variable}

\item{percent_ci}{what percent confidence do you want (e.g. .95)}

\item{lower_ci}{what is the lower bound on your interval?}

\item{upper_ci}{what is the upper bound on your interval?}

\item{beta_grid_left}{what is the lowest beta value for your grid?}

\item{beta_grid_right}{what is the highest beta value for your grid?}

\item{beta_increment}{how spaced apart do you want each beta grid element?}

\item{plot}{whether a plot should be given as a side effect}
}
\description{
This function takes some analyst specifications and returns the
"best" parameter tuple that satisfies your criteria. The analyst
specifies an ideal average for the value, and an interval (along
with a given confidence percentage). If there are tuples that
possess, both, coverage probabilities below and above the given
percentage, this function will return the tuple that a.) yields the
given mean, and b.) that is closest to the desired percentage. If there
is no "crossing" then NULL is returned. Optionally, this function
will plot the coverage probabilities.
}
\examples{
smart_pairs(fake_df, 100)
}
\keyword{elicit}
\keyword{elicitation}
\keyword{inverseGamma}
\keyword{prior}
