% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_funcs.R
\name{genAssigntmentCode}
\alias{genAssigntmentCode}
\title{A function that produces raw assignment code for a given object.}
\usage{
genAssigntmentCode(d, dest, objectName = NULL)
}
\arguments{
\item{d}{the object you have sitting in memory}

\item{dest}{the file path destination that the code will get written to}

\item{objectName}{the desired name of the object produced by the code.
Defaults to NULL, which means the name will get generated from how you
have the variable stored in your interactive session.}
}
\description{
A function that produces raw assignment code that you can copy/paste
into a script. The idea is to make it more reproducible.
}
\keyword{code}
\keyword{generation}
