% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_funcs.R
\name{log.sum.exp}
\alias{log.sum.exp}
\title{A log-sum-exp function.}
\usage{
\method{log}{sum.exp}(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\description{
This function calculates the function log(sum(exp(vec))) in a numerically stable way.
}
\examples{
log.sum.exp(rep(1e5, 3))
log(sum(exp(rep(1e5,3))))
}
\keyword{log-sum-exp}
