% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_funcs.R
\name{logLogisticRegCndtlLike}
\alias{logLogisticRegCndtlLike}
\title{A function that evaluates the log-likelihood for a logistic regression model.}
\usage{
logLogisticRegCndtlLike(y, linkedMeans)
}
\arguments{
\item{y}{observed dependent variables, assumed to be coded as {0,1}.}

\item{linkedMeans}{a numeric vector. The model matrix times the beta vector.}
}
\description{
This function evaluates the log-likelihood of a logistic regression model
in a numerically stable way. It uses the log-sum-exp trick.
}
\examples{
xiB <- rnorm(length(y), mean = 100)
logLogisticRegCndtlLike(y, xiB)
sum(dbinom(y, 1, inv.logit(xiB), TRUE)) #-Inf
}
\keyword{elicit}
\keyword{elicitation}
\keyword{inverseGamma}
\keyword{prior}
