% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_funcs.R
\name{plotSurface}
\alias{plotSurface}
\title{A function that produces a 3D plot.}
\usage{
plotSurface(lowerFirst, upperFirst, lowerSecond, upperSecond,
  numGridPointsOnEachAxis, f, contour = F, ...)
}
\arguments{
\item{lowerFirst}{lower bound of x axis}

\item{upperFirst}{upper bound of x axis}

\item{lowerSecond}{lower bound of y axis}

\item{upperSecond}{upper bound of y axis}

\item{numGridPointsOnEachAxis}{how many grid points do you want on each axis}

\item{f}{the function that takes two scalar arguments (x and y) and produces one scalar argument (z)}

\item{contour}{do you want a contour plot? (True or False)}

\item{...}{extra arguments to be passed to graphics::contour() or graphics::persp() (depending on what contour arg was set to)}
}
\description{
This function produces a 3D plot. x,y are the independent variables,
and the z-axis is the dependent.
}
\examples{
plotSurface(-50, 50, 0.0001, 50, 20, eval_log_unnormalized_posterior, F, theta=-120, zlab = "log unnorm dens", xlab = "mu", ylab = "ss")
}
\keyword{3-D}
\keyword{3D}
\keyword{3d}
\keyword{plotting}
