% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_funcs.R
\name{rhats}
\alias{rhats}
\title{A Gelman-Rubin convergence diagnostic function.}
\usage{
rhats(chain_files, burn = 0, ...)
}
\arguments{
\item{chain_files}{vector of string paths for each file of samples}

\item{burn}{the number of rows you want to discard as burn in. Defaults to 0.}

\item{...}{arguments to be passed in to lapply}
}
\description{
This function calculates the Gelman-Rubin convergence diagnostic (Rhat) on some samples.
It assumes each chain is in a separate csv file, and each file contains the samples for
all parameters. For each chain, this function splits each column into two, which means
"m" comes out to be twice the number of csv files. Finally, this function assumes each
csv has a one-line header.
}
\examples{
files <- c('/fake/path/samps1.csv', '/fake/path/samps2.csv')
rhats(files, burn = 300)
}
\keyword{Gelman-Rubin}
\keyword{rhat}
