% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_funcs.R
\name{runningMeans}
\alias{runningMeans}
\title{A running mean function.}
\usage{
runningMeans(df, cols, burn = 0)
}
\arguments{
\item{df}{A data frame of samples. Rows correspond with iterations, and columns with parameters.}

\item{cols}{column indices for the columns you want to look at.}

\item{burn}{Number of rows you want to discard as burn in. Defaults to 0.}
}
\description{
This function calculates a running mean for columns of a data frame.
}
\examples{
fake_samples <- data.frame(rnorm(100))
plot(runningMeans(fake_samples, 1), type ="l")
}
\keyword{mean}
\keyword{means}
\keyword{running}
