% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_funcs.R
\name{smart_pairs}
\alias{smart_pairs}
\title{A function that removes rows before displaying pairwise scatterplots.}
\usage{
smart_pairs(df, every)
}
\arguments{
\item{df}{a data frame of samples}

\item{every}{take every "every"rd row}
}
\description{
If pairs is choking on too much data, this function will remove rows,
and then call it. It will also plot two different colors; one color will
be given to the first half of all the samples, and the other color will
be given to the second half of all the samples.
}
\examples{
smart_pairs(fake_df, 100)
}
\keyword{pairs}
\keyword{pairwise}
\keyword{scatterplot}
